rename to pebble_digraph
